+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changelog
 1.11 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Use `self.context.wl_isLocked()` to show locking icon so it does not break
+  on non lockable objects.
+  [gbastien]
 
 1.10 (2018-01-06)
 -----------------